Create Protocol interface
