+++ src/main/java/com/pwdd/server/protocol/Protocol.java
@@ -1,0 +1,14 @@
+package com.pwdd.server.protocol;
+
+import com.pwdd.server.responders.IResponder;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+
+public interface Protocol {
+  String version = "HTTP/1.1";
+
+  byte[] processResponse(BufferedReader request) throws IOException;
+
+  IResponder[] responders();
+}
