+++ src/main/java/com/pwdd/server/server/ConnectionManager.java
@@ -1,8 +1,7 @@
 package com.pwdd.server.server;
 
-import com.pwdd.server.RequestParser;
-import com.pwdd.server.responders.ResponseBuilder;
-import com.pwdd.server.utils.FileHandler;
+import com.pwdd.server.responders.*;
+import com.pwdd.server.protocol.GET;
 
 import java.io.*;
 import java.net.Socket;
@@ -10,11 +9,9 @@ import java.net.Socket;
 class ConnectionManager implements Runnable {
   private File rootDirectory;
   private final Socket socket;
-  private final ResponseBuilder responseBuilder;
 
-  ConnectionManager(Socket _socket, ResponseBuilder _responseBuilder, File _rootDirectory) {
+  ConnectionManager(Socket _socket, File _rootDirectory) {
     this.socket = _socket;
-    this.responseBuilder = _responseBuilder;
     this.rootDirectory = _rootDirectory;
   }
 
@@ -31,21 +28,11 @@ class ConnectionManager implements Runnable {
   @Override
   public void run() {
     try {
-      String uri = RequestParser.header(getRequestFrom(socket)).get("URI");
-      File file = defineFile(uri);
-      byte[] response = this.responseBuilder.response(file);
+      byte[] response = new GET(rootDirectory).processResponse(getRequestFrom(socket));
       sendResponseTo(socket, response);
       socket.close();
     } catch (Exception e) {
       e.printStackTrace();
-    }
-  }
-
-  private File defineFile(String uri) {
-    if (uri.toLowerCase().equals("/hello") || uri.toLowerCase().equals("/hello/")) {
-      return new File("/hello");
-    } else {
-      return FileHandler.uriToFile(rootDirectory.getAbsolutePath(), uri);
     }
   }
 }
