+++ src/main/java/com/pwdd/server/server/Server.java
@@ -1,7 +1,5 @@
 package com.pwdd.server.server;
 
-import com.pwdd.server.responders.ResponseBuilder;
-
 import java.io.File;
 import java.net.ServerSocket;
 import java.net.Socket;
@@ -12,12 +10,10 @@ class Server implements Runnable {
   private final int portNumber;
   private Boolean listening = false;
   private ConnectionManager connectionManager;
-  private final ResponseBuilder responseBuilder;
   private File rootDirectory;
 
-  Server(int _portNumber, ResponseBuilder _responseBuilder, File _rootDirectory) {
+  Server(int _portNumber, File _rootDirectory) {
     this.portNumber = _portNumber;
-    this.responseBuilder = _responseBuilder;
     this.rootDirectory = _rootDirectory;
   }
 
@@ -30,7 +26,7 @@ class Server implements Runnable {
   }
 
   private void startConnectionHandler() {
-    connectionManager = new ConnectionManager(socket, responseBuilder, rootDirectory);
+    connectionManager = new ConnectionManager(socket, rootDirectory);
   }
 
   @Override
